Fix LoaderRing in Submit buttons in Form
